% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Anderson_Darling_test}
\alias{Anderson_Darling_test}
\title{Anderson-Darling normality test}
\usage{
Anderson_Darling_test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with three elements
\enumerate{
\item is_normality: logical value. TRUE: x is normal distribution
\item statistic: the Anderson-Darling test statistic value.
The higher the value, the lower the probability of normality.
\item pval: the p-value of the statistics. pval > 0.05 indicates normal distribution.
}
}
\description{
Test the data normality (normal distribution)
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Anderson_Darling_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#> [1] 0.1602067
#> $pval
#> [1] 0.9470751

#' x <- stats::runif(100)  # This is non-normal distribution
Anderson_Darling_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#> [1] 1.14719
#> $pval
#> [1] 0.0050856
}
\references{
D’Agostino, RalphB., 2017.
Goodness-of-Fit Techniques.
1st ed. Routledge. pg. 127 (Table 4.9).
https://doi.org/10.1201/9780203753064
}
