% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned_rank_transform.R
\name{art_1}
\alias{art_1}
\title{One-way aligned rank transform ANOVA}
\usage{
art_1(data, formula)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{A formula specifying the model, for example, yield ~ fertilizer_level.}
}
\value{
ANOVA (type 3) result
}
\description{
One-way aligned rank transform ANOVA
}
\examples{
set.seed(1)
df0 <- data.frame(
    group = as.factor(rep(c("A", "B", "C"), each = 10)),
    norm_data = c(rnorm(10, -1, 2), rnorm(10, 3, 2), rnorm(10, 0, 1.5)),
    skew_data = c(sqrt(rnorm(10, -7.2, 2) ^ 2), runif(10), runif(10))
)
art_1(df0, skew_data ~ group)
#> Anova Table (Type 3)
#> Response: group
#> Sum Sq Df F value    Pr(>F)
#> (Intercept) 6502.5  1 263.022 1.917e-15 ***
#> x           1580.0  2  31.955 7.623e-08 ***
#> Residuals    667.5 27
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
}
