% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoscedasticity.R
\name{barlett_test}
\alias{barlett_test}
\title{Barlett's test}
\usage{
barlett_test(data, formula)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula}
}
\value{
returns a list containing three elements:
\enumerate{
\item is_var_equal
\item statistic
\item pval
}
}
\description{
This is a wrapper for the \code{stats::bartlett.test()} function.
Computes Barlett's test for homogeneity of variance across groups (homoscedasticity).
}
\examples{
utils::data("iris", package = "datasets")
barlett_test(iris, Petal.Length ~ Species)
#> $is_var_equal
#> [1] FALSE
#>
#> $statistic
#> Bartlett's K-squared
#>              55.4225
#>
#> $pval
#> [1] 2.005419e-08

}
\seealso{
\code{\link[stats:bartlett.test]{stats::bartlett.test()}}
}
\author{
Joon-Keat Lai
}
