% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_letter_display.R
\name{compact_letter_display}
\alias{compact_letter_display}
\title{Compact Letter Display (CLD)}
\usage{
compact_letter_display(
  groups,
  means,
  comparisons,
  pval,
  alpha = 0.05,
  comparison_symbol = " |vs| ",
  display_symbols = base::letters,
  descending = TRUE
)
}
\arguments{
\item{groups}{A character vector with the names of each groups (i.e. independent variables).}

\item{means}{A numeric vector with the mean values of each groups (i.e. response variable).}

\item{comparisons}{A character vector with the comparisons annotation, yielded from post-hoc test.}

\item{pval}{A numeric vector with the p-values corresponding to the \code{comparisons}.}

\item{alpha}{Default: 0.05; a numeric scalar.}

\item{comparison_symbol}{Default: " |vs| "}

\item{display_symbols}{Default: c("a", "b", "c", ...)}

\item{descending}{Default: TRUE}
}
\value{
A character named-vector
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\examples{
utils::data("iris", package = "datasets")
stats_desc <- tapply(iris$Sepal.Length, iris$Species, "mean")
stats_res <- stats::TukeyHSD(aov(Sepal.Length ~ Species, iris))
stats_res <- as.data.frame(stats_res$Species)
compact_letter_display(
    groups = names(stats_desc),
    means = stats_desc,
    comparisons = row.names(stats_res),
    pval = stats_res$`p adj`,
    comparison_symbol = "-"
)
#> setosa versicolor  virginica
#>   "c"      "b"        "a"
}
\references{
Piepho, H.P., 2004.
An Algorithm for a Letter-Based Representation of All-Pairwise Comparisons.
Journal of Computational and Graphical Statistics 13, 456–466.
https://doi.org/10.1198/1061860043515
}
\author{
Joon-Keat Lai
}
