% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{could_be_number}
\alias{could_be_number}
\title{Check if the element in character strings could be numeric}
\usage{
could_be_number(x)
}
\arguments{
\item{x}{Character strings}
}
\value{
Boolean values
}
\description{
Check if the element in character strings could be numeric
}
\examples{
x <- c("1a", ".7", "46", "2.3.3", "1.2r", "1.2", "1e4", "a1", "5L", "-.22", -Inf, NA, NaN)
could_be_number(x)
#>    1a    .7    46 2.3.3  1.2r   1.2   1e4    a1    5L  -.22  -Inf    NA   NaN
#> FALSE  TRUE  TRUE FALSE FALSE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE FALSE FALSE
}
