% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Cramer_von_Mises_test}
\alias{Cramer_von_Mises_test}
\title{Cramer-von Mises normality test}
\usage{
Cramer_von_Mises_test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with three elements:
\enumerate{
\item is_normality: logical value. TRUE indicates x is a normal distribution.
\item statistic: the Cramer-von Mises test statistic value (W).
The higher the value, the lower the probability of normality.
\item pval: the p-value for the test. pval > 0.05 indicates normal distribution.
}
}
\description{
The Cramer-von Mises test is an empirical distribution function (EDF) omnibus test
for the hypothesis of normality. The test statistic is
\deqn{
W = \frac{1}{12 n} + \sum_{i=1}^{n} \left(p_{(i)} - \frac{2i-1}{2n}\right)^2,
}{W = 1/(12n)  +  \sum_{i=1}^n (p_(i) - (2i-1)/(2n))^2,} \
where \eqn{p_{(i)} = \Phi([x_{(i)} - \overline{x}]/s)},
\eqn{\Phi} is the cumulative distribution function of the standard normal distribution,
\eqn{\overline{x}} and \eqn{s} are the mean and standard deviation.
The p-value is computed from the modified statistic \eqn{Z = W (1.0 + 0.5/n)}
according to Table 4.9 in Stephens (1986).
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Cramer_von_Mises_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#>         W
#> 0.0260313
#> $pval
#> [1] 0.8945191

x <- stats::runif(100)  # This is non-normal distribution
Cramer_von_Mises_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#>         W
#> 0.2746561
#> $pval
#> [1] 0.0006342302

}
\references{
Stephens, M.A. 1986.
Tests based on EDF statistics.
Eds.: D'Agostino, R.B. and Stephens, M.A.
Goodness-of-Fit Techniques.
Marcel Dekker, New York.

Thode Jr., H.C. 2002.
Testing for  Normality.
Marcel Dekker, New York.
}
\seealso{
\code{\link[nortest:cvm.test]{nortest::cvm.test()}}
}
\author{
Joon-Keat Lai
}
