% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataframe_to_list}
\alias{dataframe_to_list}
\title{Convert dataframe to list}
\usage{
dataframe_to_list(data, formula)
}
\arguments{
\item{data}{A dataframe.}

\item{formula}{A formula specifying the value and key names. For example, \code{val ~ group}.}
}
\value{
A list
}
\description{
Convert dataframe to list
}
\examples{
df0 <- data.frame(group = rep(c("A", "B", "C"), 5), val = rep(1:3, 5))
dataframe_to_list(df0, val ~ group)
#> $A
#> [1] 1 1 1 1 1
#>
#> $B
#> [1] 2 2 2 2 2
#>
#> $C
#> [1] 3 3 3 3 3
}
