% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dunnett_test.R
\name{Dunnett_test}
\alias{Dunnett_test}
\title{Dunnett's test}
\usage{
Dunnett_test(
  data,
  formula,
  control,
  alpha = 0.05,
  p_adjust_method = "none",
  nsim = 1e+05,
  seed = 1,
  descending = NULL
)
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{control}{The which group as the control.}

\item{alpha}{Default: 0.05}

\item{p_adjust_method}{A character string (default: "none"). Other options: \code{stats::p.adjust.methods}.}

\item{nsim}{Number of times of simulation.}

\item{seed}{Random seed.}

\item{descending}{Sort the group by their mean value.}
}
\value{
A list with four vectors.
\enumerate{
\item tests: A message showing the statistical methods applied on the dataset.
\item pre_hoc: The result of pmnibus test.
\item post_hoc: includes statistics parameters for each pairwise comparisons.
\item cld: A dataframe reporting the descriptive stats and compact letter display.
}
}
\description{
Comparing multiple groups with a fixed control group. This could be more sensitive
to show significance than the pairwise comparison approaches.
}
\examples{
set.seed(1)
df0 <- data.frame(
    fertilizer = rep(c("A", "B", "C", "D", "E"), each = 10),
    yield = c(
        stats::rnorm(10, 8, 1.2),
        stats::rnorm(10, 9.1, 1.1),
        stats::rnorm(10, 8.5, 1.3),
        stats::rnorm(10, 10, 1.05),
        stats::rnorm(10, 7, 1.17))
)
out <- Dunnett_test(df0, yield ~ fertilizer, control = "A")
out$cld
#>   GROUP  N       AVG        SD      MED      MIN       MAX CLD
#> 1     A 10  8.158643 0.9367031 8.307891 6.997246  9.914337
#> 2     B 10  9.373729 1.1764663 9.641060 6.663830 10.762959   *
#> 3     C 10  8.326225 1.2422899 8.511984 5.913843  9.694671  ns
#> 4     D 10 10.126767 0.8489929 9.940613 8.554087 11.426614 ***
#> 5     E 10  7.156940 0.6970463 7.147558 6.172231  8.030896  ns
}
