% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_cld_pos}
\alias{estimate_cld_pos}
\title{Estimate the suitable compact letter y-position used in the ggplot boxplot}
\usage{
estimate_cld_pos(x)
}
\arguments{
\item{x}{The maximum values of each groups}
}
\value{
numeric values (y-axis position)
}
\description{
Estimate the suitable compact letter y-position used in the ggplot boxplot
}
\examples{
library(ggplot2)
data("iris")
res <- statool::oneway_test(iris, Sepal.Length ~ Species)
cld <- res$result
cld$y_pos <- estimate_cld_pos(cld$MAX)
print(cld)
#>                GROUPS  N   AVG        SD MED MIN MAX CLD y_pos
#> virginica   virginica 50 6.588 0.6358796 6.5 4.9 7.9   a 9.286
#> versicolor versicolor 50 5.936 0.5161711 5.9 4.9 7.0   b 8.386
#> setosa         setosa 50 5.006 0.3524897 5.0 4.3 5.8   c 7.186

ggplot(iris, aes(Species, Sepal.Length, color = Species)) +
    geom_boxplot() +
    geom_point() +
    geom_text(
        inherit.aes = FALSE,
        data = cld,
        mapping = aes(GROUPS, y_pos, label = CLD)
    )
}
