% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoscedasticity.R
\name{fligner_test}
\alias{fligner_test}
\title{Fligner's test}
\usage{
fligner_test(data, formula)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula}
}
\value{
returns a list containing three elements:
\enumerate{
\item is_var_equal
\item statistic
\item pval
}
}
\description{
This is a wrapper for the \code{stats::fligner.test()} function.
Computes Fligner's test for homogeneity of variance across groups (homoscedasticity).
}
\examples{
utils::data("iris", package = "datasets")
fligner_test(iris, Petal.Length ~ Species)
#> $is_var_equal
#> [1] FALSE
#>
#> $statistic
#> Fligner-Killeen:med chi-squared
#>                        34.70802
#>
#> $pval
#> [1] 2.90569e-08

}
\seealso{
\code{\link[stats:fligner.test]{stats::fligner.test()}}
}
\author{
Joon-Keat Lai
}
