% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{Grubbs_test}
\alias{Grubbs_test}
\title{Grubbs's test}
\usage{
Grubbs_test(x, xs = NULL, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector (normal distribution) to test for an outlier.}

\item{xs}{Which observation is suspected as an outlier.
If this is specified, then one-tailed test is performed. If NULL, then performed two-tailed test.}

\item{alpha}{Default: 0.05}
}
\value{
A list containing 3 vectors:
\enumerate{
\item G: The statistic value for the test.
\item Gcrit: The threshold of the test to reject null hypothesis.
\item is_outlier: Boolean values indicating which one is the outlier (G > Gcrit).
}
}
\description{
Description
}
\examples{
set.seed(1)
x <- c(round(rnorm(10), 2), 5)
Grubbs_test(x)
#> $G
#> [1] 2.689996
#>
#> $Gcrit
#> [1] 2.35473
#>
#> $is_outlier
#> -0.63  0.18 -0.84   1.6  0.33 -0.82  0.49  0.74  0.58 -0.31     5
#> FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
}
