% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_normality}
\alias{is_normality}
\title{Check data normality}
\usage{
is_normality(data, formula, alpha = 0.05)
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}
}
\value{
A boolean value indicating the acceptance of normality assumption
}
\description{
Check data normality
}
