% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{is_outlier}
\alias{is_outlier}
\title{Title}
\usage{
is_outlier(
  x,
  method = "iqr",
  n.iqr = 1.5,
  n.sd = 3,
  trim = 0.1,
  lower_quantile = 0.2,
  upper_quantile = 0.8,
  use_median = TRUE
)
}
\arguments{
\item{x}{A numeric vector}

\item{method}{Detect the outliers with either standard deviation (sd), interquantile range (iqr), or mean (mean)}

\item{n.iqr}{Default: 1.5}

\item{n.sd}{Default: 3}

\item{trim}{Default: 0.1}

\item{lower_quantile}{Default: 0.2}

\item{upper_quantile}{Default: 0.8}

\item{use_median}{Default: TRUE}
}
\value{
A boolean vector
}
\description{
Title
}
\examples{
set.seed(1)
x1 <- c(rnorm(5), 4)
is_outlier(x1, method = "sd", n.sd = 2)
#> FALSE FALSE FALSE FALSE FALSE  TRUE
}
