% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoscedasticity.R
\name{is_var_equal}
\alias{is_var_equal}
\title{Homoscedasticity}
\usage{
is_var_equal(data, formula, alpha = 0.05, method = "median")
}
\arguments{
\item{data}{data frame}

\item{formula}{formula}

\item{alpha}{(default: 0.05)}

\item{method}{Available options are: "median" (default), "mean", and "trim_mean".}
}
\value{
Logical value
}
\description{
Test if the variances among groups are equal.
}
\examples{
utils::data("iris", package = "datasets")
is_var_equal(iris, Petal.Length ~ Species)
#> [1] FALSE
}
