% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_vector}
\alias{is_vector}
\title{Check vector
Check if an object is a vector.}
\usage{
is_vector(x)
}
\arguments{
\item{x}{An object}
}
\value{
Boolean value
}
\description{
Check vector
Check if an object is a vector.
}
\examples{
is_vector(c(1, 2, 3))
#> TRUE

is_vector(list(a = c(1, 2, 3), b = c("a", "b", "c")))
#> FALSE
}
