% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Jarque_Bera_test}
\alias{Jarque_Bera_test}
\title{Jarque-Bera normality test}
\usage{
Jarque_Bera_test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with three elements
\enumerate{
\item is_normality: logical value. TRUE: x is normal distribution
\item statistic: the Jarque-Bera test statistic value.
The higher the value, the lower the probability of normality.
\item pval: the p-value of the statistics. pval >= 0.05 indicates normal distribution.
}
}
\description{
Test the data normality (normal distribution).
Insensitive to uniform distribution. Not recommended.
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Jarque_Bera_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#> [1] 0.1103686
#> $pval
#> [1] 0.9463107

set.seed(1)
x <- stats::rchisq(100, 1)  # This is non-normal distribution
Jarque_Bera_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#> [1] 615.3756
#> $pval
#> [1] 2.359915e-134

}
\references{
Jarque, C.M. and Bera, A.K. 1980.
Efficient tests for normality, homoscedasticity and serial independence of regression residuals.
Economics letters, 6(3), 255-259.
}
\author{
Joon-Keat Lai
}
