% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis of the sample / population distribution}
\usage{
kurtosis(x, .population = FALSE)
}
\arguments{
\item{x}{A vector of numeric values. NA values will be discarded automatically.}

\item{.population}{Calculate the kurtosis of the population or the sample.}
}
\value{
Numeric value
}
\description{
Return the kurtosis of the sample / population distribution.
}
\examples{
set.seed(1)
x <- stats::rnorm(10, 10)
kurtosis(x)
#> -0.3169031
}
