% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoscedasticity.R
\name{levene_test}
\alias{levene_test}
\title{Levene's test}
\usage{
levene_test(data, formula, method = NULL)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula}

\item{method}{options are: "trim_mean" (default), "mean", "median"}
}
\value{
returns a list containing three elements:
\enumerate{
\item is_var_equal
\item statistic
\item pval
}
}
\description{
Computes Levene's test for homogeneity of variance across groups (homoscedasticity).
}
\examples{
utils::data("iris", package = "datasets")
levene_test(iris, Petal.Length ~ Species)
#> $is_var_equal
#> [1] FALSE
#> $statistic
#> Analysis of Variance Table
#>
#> Response: diff
#>            Df Sum Sq Mean Sq F value    Pr(>F)
#> x           2 2.6526 1.32629  20.045 2.005e-08 ***
#> Residuals 147 9.7265 0.06617
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>
#> $pval
#> [1] 2.005419e-08

}
\references{
Fox J. and Weisberg S. 2019.
\emph{An R Companion to Applied Regression},
Third Edition, Sage.
}
\seealso{
\code{\link[car:leveneTest]{car::leveneTest()}}
}
\author{
Joon-Keat Lai
}
