% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Lilliefors_test}
\alias{Lilliefors_test}
\title{Lilliefors normality test}
\usage{
Lilliefors_test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with three elements:
\enumerate{
\item is_normality: logical value. TRUE indicates x is a normal distribution.
\item statistic: the Lilliefors test statistic value (D).
The higher the value, the lower the probability of normality.
\item pval: the p-value for the test. pval > 0.05 indicates normal distribution.
}
}
\description{
The Lilliefors test, which is a modified version of the Kolmogoro-Smirnov test,
is an empirical distribution function (EDF) omnibus test for the hypothesis of
normality.
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Lilliefors_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#>          D
#> 0.04701381
#> $pval
#> [1] 0.8479244

x <- stats::runif(100)  # This is non-normal distribution
Lilliefors_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#>          D
#> 0.1007912
#> $pval
#> [1] 0.01400096

}
\references{
Dallal G.E. and Wilkinson L. 1986.
An analytic approximation to the distribution of Lilliefors' test for normality.
The American Statistician. 40:294-296.
}
\seealso{
\code{\link[nortest:lillie.test]{nortest::lillie.test()}}
}
\author{
Joon-Keat Lai
}
