% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scaling.R
\name{mean_scaling}
\alias{mean_scaling}
\title{Mean scaling}
\usage{
mean_scaling(x, range = c(0, 1))
}
\arguments{
\item{x}{A numeric vector.}

\item{range}{A numeric vector with length of 2. The desired range for the data normalization.
The default is c(0, 1).}
}
\value{
A numeric vector, i.e., the normalized data.
}
\description{
Feature scaling method normalizes the input data to a desired range (generally from 0 to 1).
The transformation was done by:
\verb{[(x - mean(x)) / (max(x) - min(x))] * (range_ceiling - range_floor) + range_floor}
}
\examples{
set.seed(1)
x <- stats::rnorm(10)
mean_scaling(x)
#> [1] -0.31208756  0.02116103 -0.39813552  0.60186448  0.08116509
#> [6] -0.39189908  0.14612896  0.24933958  0.18247433 -0.18001130

mean_scaling(x, range = c(-2, 2))
#> [1] -3.2483502 -1.9153559 -3.5925421  0.4074579 -1.6753396
#> [6] -3.5675963 -1.4154841 -1.0026417 -1.2701027 -2.7200452
}
\author{
Joon-Keat Lai
}
