% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pval2asterisk}
\alias{pval2asterisk}
\title{Assign asterisks according to the p-value}
\usage{
pval2asterisk(
  x,
  cutpoints = c(0, 0.001, 0.01, 0.05, 1),
  symbols = c("***", "**", "*", "ns"),
  superscript = FALSE
)
}
\arguments{
\item{x}{A numeric vector which values are ranged from 0 to 1 (the p-values).}

\item{cutpoints}{Default: c(0, 0.001, 0.01, 0.05, 1).}

\item{symbols}{Default: c("\emph{\strong{", "}", "}", "ns").}

\item{superscript}{Show the asterisks as superscript.}
}
\value{
A vector of asterisks (\code{*}) or \code{ns} symbols
}
\description{
Assign asterisks according to the p-value
}
\examples{
pvals <- c(0.1, 0.05, 0.001, 0.006, 0.00009, 0.5)
ret <- pval2asterisk(pvals)
ret
#> 0.1  0.05 0.001 0.006 9e-05   0.5
#>  ns     *   ***    **   ***    ns
}
\author{
Joon-Keat Lai
}
