% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REGWQ_test.R
\name{REGWQ_test}
\alias{REGWQ_test}
\title{Ryan, Einot, Gabriel, Welsh Studentized Range Q (REGWQ) test}
\usage{
REGWQ_test(data, formula, alpha = 0.05, p_adjust_method = "none")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Options please refer to \code{stats::p.adjust.methods}.}
}
\value{
A list with four elements:
\enumerate{
\item tests: A message showing the statistical methods applied on the dataset.
\item pre_hoc: The result of pmnibus test.
\item post_hoc: includes statistics parameters for each pairwise comparisons.
\item cld: A dataframe reporting the descriptive stats and compact letter display.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\examples{
set.seed(1)
group <- rep(c("A", "B", "C", "D", "E"), each = 10)
val <- c(rnorm(10, 1, 1.5), rnorm(10, 2, 1), rnorm(10, -4, 2), rnorm(10), rnorm(10))
df0 <- data.frame(group = group, val = val)
out <- REGWQ_test(df0, val ~ group)
out$cld
#>   GROUP  N        AVG        SD         MED        MIN        MAX CLD
#> 1     B 10  2.2488450 1.0695148  2.49187228 -0.2146999  3.5117812   a
#> 2     A 10  1.1983042 1.1708789  1.38486332 -0.2534429  3.3929212  ab
#> 3     E 10  0.1341367 0.5957661  0.12611787 -0.7074952  0.8811077   b
#> 4     D 10  0.1207302 0.8085646 -0.05655922 -1.3770596  1.3586796   b
#> 5     C 10 -4.2673465 1.9112152 -3.98156376 -7.9787034 -2.1620453   c
}
\references{
Howell, D.C. 2013.
Statistical methods for psychology (8th ed.). pg. 393.
Wadsworth Cengage Learning, Belmont, CA.
}
\author{
Joon-Keat Lai
}
