% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scaling.R
\name{robust_scaling}
\alias{robust_scaling}
\title{Robust scaling}
\usage{
robust_scaling(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector, i.e., the normalized data.
}
\description{
Feature scaling method normalizes the input data by the
median and the inter-quantile range (IQR, P75 - P25).
The transformation was done by:
\code{(x - median(x)) / IQR(x)}
}
\examples{
set.seed(1)
x <- stats::rnorm(10)
robust_scaling(x)
#> [1] -0.80284101 -0.06630921 -0.99302054  1.21713675  0.06630921
#> [6] -0.97923702  0.20988958  0.43800127  0.29021856 -0.51093170
}
\author{
Joon-Keat Lai
}
