% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Ryan_Joiner_test}
\alias{Ryan_Joiner_test}
\title{Ryan-Joiner normality test}
\usage{
Ryan_Joiner_test(x, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector}

\item{alpha}{The error tolerance, either 0.1, 0.05 (default), or 0.01.}
}
\value{
A list with three elements
\enumerate{
\item is_normality: logical value. TRUE: x is normal distribution
\item statistic: the Ryan-Joiner test statistic (Rp) and its critical value (Rc).
Rp > Rc accepts normality assumption.
\item pval: return an \code{NA}. the p-value of the statistics is currently not available.
}
}
\description{
Test the data normality (normal distribution).
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Ryan_Joiner_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#>        Rp           Rc
#> 0.9979065    0.9874371
#> $pval
#> [1] NA

set.seed(1)
x <- stats::runif(100, 1)  # This is non-normal distribution
Ryan_Joiner_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#>        Rp            Rc
#>      -Inf     0.9874371
#> $pval
#> [1] NA

}
\references{
Ryan T.A. and Joiner B.L. 1976.
Normal probability plots and tests for normality.
Technical Report.
Statistics Department, The Pennsylvania State University.
}
