% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd_population}
\alias{sd_population}
\title{Standard deviation of the population}
\usage{
sd_population(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
standard deviation value for the population
}
\description{
Standard deviation of the population
}
\examples{
set.seed(1)
sd_population(rnorm(100))
#> 0.8936971
}
