% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Shapiro_Francia_test}
\alias{Shapiro_Francia_test}
\title{Shapiro-Francia normality test}
\usage{
Shapiro_Francia_test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with three elements:
\enumerate{
\item is_normality: logical value. TRUE indicates x is a normal distribution.
\item statistic: the Shapiro-Francia test statistic value (W).
The higher the value, the lower the probability of normality.
\item pval: the p-value for the test. pval > 0.05 indicates normal distribution.
}
}
\description{
The Shapiro-Francia test is a type of regression and correlation test for normality
evaluation. It estimates the squared correlation between the ordered sample values
and the approximate expected ordered quantiles from the standard normal distribution.
The p-value is computed from the formula given by Royston (1993).
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Shapiro_Francia_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#>         W
#> 0.9958174
#> $pval
#> [1] 0.9718829

x <- stats::runif(100)  # This is non-normal distribution
Shapiro_Francia_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#>         W
#> 0.9386237
#> $pval
#> [1] 0.0003371829

}
\references{
Royston, P. 1993.
A pocket-calculator algorithm for the Shapiro-Francia test for non-normality: an application to medicine.
Statistics in Medicine. 12, 181-184.

Thode Jr., H.C. 2002.
Testing for  Normality.
Marcel Dekker, New York.
}
\seealso{
\code{\link[nortest:sf.test]{nortest::sf.test()}}
}
\author{
Joon-Keat Lai
}
