% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{skewness}
\alias{skewness}
\title{Skewness of the sample / population distribution}
\usage{
skewness(x, .population = FALSE)
}
\arguments{
\item{x}{A vector of numeric values. NA values will be discarded automatically.}

\item{.population}{Calculate the skewness of the population or the sample.}
}
\value{
Numeric value
}
\description{
Return the skewness of the sample / population distribution.
The skewness value for a normal distribution should be close to zero.
A negative skewness value indicates left-skewed distribution (concentrated on the right);
A positive skewness value indicates right-skewed distribution (concentrated on the left).
}
\examples{
set.seed(1)
x <- stats::rnorm(10, 10)
skewness(x)
#> 0.3512426
}
