% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_is_number}
\alias{string_is_number}
\title{Check if the character could be numeric
Check if the character could be numeric}
\usage{
string_is_number(
  x,
  include_L_sign = TRUE,
  exclude_infinite = TRUE,
  exclude_boolean = TRUE
)
}
\arguments{
\item{x}{A character vector}

\item{include_L_sign}{(Default: TRUE). If TRUE, for example, "5L" will be considered as a number.}

\item{exclude_infinite}{(Default: TRUE). If TRUE, for example, "Inf" will NOT be considered as a number.}

\item{exclude_boolean}{(Default: TRUE). If TRUE, for example, "TRUE" will NOT be considered as a number.}
}
\value{
A boolean vector
}
\description{
Check if the character could be numeric
Check if the character could be numeric
}
\examples{
x <- c("1a", ".7", "46", "2.3.3", "1.2r", "1.2", "1e4", "a1", "5L", "-.22", -Inf, NA, NaN)
string_is_number(x)
#>    1a    .7    46 2.3.3  1.2r   1.2   1e4    a1    5L  -.22  -Inf    NA   NaN
#> FALSE  TRUE  TRUE FALSE FALSE  TRUE  TRUE FALSE  TRUE  TRUE  FALSE FALSE FALSE
}
