% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Student's t-Test}
\usage{
t_test(
  data,
  formula = NULL,
  y1 = NULL,
  y2 = NULL,
  alternative = c("two.sided", "less", "greater"),
  var_equal = NULL,
  alpha = 0.05,
  paired = FALSE
)
}
\arguments{
\item{data}{A data.frame}

\item{formula}{A formula y ~ x specifying the response and independant variable.}

\item{y1}{A numeric vector of data values}

\item{y2}{A numeric vectoe of another data values}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{var_equal}{A logical variable indicating whether to treat the two variances as being equal.
If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{alpha}{Confidence level (1 - alpha) of the interval.}

\item{paired}{A logical indicating whether you want a paired t-test.}
}
\value{
A data.frame
}
\description{
Performs one and two sample t-tests on vectors of data.
}
\examples{
y1 <- rnorm(10)
y2 <- rnorm(10, 5, 3)
t_test(y1 = y1, y2 = y2)
}
