% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey_Kramer_test.R
\name{Tukey_Kramer_test}
\alias{Tukey_Kramer_test}
\title{Tukey-Kramer test (Modified Tukey-HSD for unequal sample size)}
\usage{
Tukey_Kramer_test(data, formula, alpha = 0.05, p_adjust_method = "none")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Other options: \code{stats::p.adjust.methods}.}
}
\value{
A list with two vectors.
\enumerate{
\item result: consists of descriptive statistics and compact letter display;
\item comparisons: includes statistics parameters for each pairwise comparisons.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\references{
Howell, D.C. 2013.
Statistical methods for psychology (8th ed.). pg. 393.
Wadsworth Cengage Learning, Belmont, CA.

Kramer, C.Y. 1956.
Extension of Multiple Range Tests to Group Means with Unequal Numbers of Replications.
Biometrics, vol. 12, no. 3, pp. 307-310. JSTOR.
https://doi.org/10.2307/3001469.
}
