% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scaling.R
\name{z_score_scaling}
\alias{z_score_scaling}
\title{Z-score scaling (Standardization)}
\usage{
z_score_scaling(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector, i.e., the standardized data.
}
\description{
Feature scaling method standardizes the input data to Z-score.
The transformation was done by:
\code{(x - mean(x)) / sd(x)}
}
\examples{
set.seed(1)
x <- stats::rnorm(10)
z_score_scaling(x)
#> [1] -0.97190653  0.06589991 -1.23987805  1.87433300  0.25276523
#> [6] -1.22045645  0.45507643  0.77649606  0.56826358 -0.56059319
}
\author{
Joon-Keat Lai
}
