% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{Anderson_Darling_test}
\alias{Anderson_Darling_test}
\title{Anderson-Darling normality test}
\usage{
Anderson_Darling_test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with three elements:
\enumerate{
\item is_normality: logical value. TRUE indicates x is a normal distribution.
\item statistic: the Anderson-Darling test statistic value.
The higher the value, the lower the probability of normality.
\item pval: the p-value for the test. pval > 0.05 indicates normal distribution.
}
}
\description{
The Anderson-Darling test is an empirical distribution function (EDF) omnibus test
for the hypothesis of normality. The test statistic is
\deqn{
A = -n -\frac{1}{n} \sum_{i=1}^{n} [2i-1]
[\ln(p_{(i)}) + \ln(1 - p_{(n-i+1)})],
}
where
\eqn{p_{(i)} = \Phi([x_{(i)} - \overline{x}]/s)}.
\eqn{\Phi} is the cumulative distribution function of the standard normal distribution;
\eqn{\overline{x}} and \eqn{s} are mean and standard deviation of the data values.
The p-value is computed from the modified statistic
\eqn{Z=A (1.0 + 0.75/n +2.25/n^{2})} according to the Table 4.9 in D’Agostino (2017).
}
\examples{
set.seed(1)
x <- stats::rnorm(100)  # This is normal distribution
Anderson_Darling_test(x)
#> $is_normality
#> [1] TRUE
#> $statistic
#> [1] 0.1602067
#> $pval
#> [1] 0.9470751

x <- stats::runif(100)  # This is non-normal distribution
Anderson_Darling_test(x)
#> $is_normality
#> [1] FALSE
#> $statistic
#> [1] 1.14719
#> $pval
#> [1] 0.0050856

}
\references{
D’Agostino, R.B. 2017.
Goodness-of-Fit Techniques.
1st ed. Routledge. pg. 127 (Table 4.9).
https://doi.org/10.1201/9780203753064
}
\seealso{
nortest::ad.test()
}
\author{
Joon-Keat Lai
}
