% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dunn_test.R
\name{Dunn_test}
\alias{Dunn_test}
\title{Dunn's test}
\usage{
Dunn_test(data, formula, alpha = 0.05, p_adjust_method = "bonferroni")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{Character string (default: "none"). Other options: \code{stats::p.adjust.methods}.}
}
\value{
A list with two vectors. \
\enumerate{
\item result: consists of descriptive statistics and compact letter display;
\item comparisons: includes statistics parameters for each pairwise comparisons.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\references{
Dunn, O.J., 1964.
Multiple Comparisons Using Rank Sums.
Technometrics 6, 241–252.
https://doi.org/10.1080/00401706.1964.10490181
}
\author{
Joon-Keat Lai
}
