% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dunn_test.R
\name{Dunn_test}
\alias{Dunn_test}
\title{Dunn's test}
\usage{
Dunn_test(data, formula, alpha = 0.05, p_adjust_method = "holm")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (Default: 0.05).}

\item{p_adjust_method}{Character string (default: "holm"). Other options: \code{stats::p.adjust.methods}.}
}
\value{
A list with two vectors. \
\enumerate{
\item result: consists of descriptive statistics and compact letter display;
\item comparisons: includes statistics parameters for each pairwise comparisons.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\examples{
set.seed(1)
df0 <- data.frame(
    group = rep(c("A", "B", "C"), each = 10),
    norm_data = c(rnorm(10, -1, 2), rnorm(10, 3, 2), rnorm(10, 0, 1.5)),
    skew_data = c(sqrt(rnorm(10, -7.2, 2) ^ 2), runif(10), runif(10))
)

out <- Dunn_test(df0, skew_data ~ group)
res <- out$result
res
#>   GROUP  N       AVG        SD       MED        MIN       MAX CLD
#> 1     A 10 6.9585396 1.6171293 7.3131184 4.48264090 9.9541191   a
#> 2     C 10 0.5674507 0.2713306 0.6236108 0.05893438 0.8762692   b
#> 3     B 10 0.4053906 0.2437230 0.3626733 0.12169192 0.7570871   b
}
\references{
Dunn, O.J., 1964.
Multiple Comparisons Using Rank Sums.
Technometrics 6, 241–252.
https://doi.org/10.1080/00401706.1964.10490181
}
\author{
Joon-Keat Lai
}
