% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dunn_test.R
\name{Dunn_test}
\alias{Dunn_test}
\title{Dunn's test}
\usage{
Dunn_test(
  data,
  formula,
  alpha = 0.05,
  p_adjust_method = "none",
  descending = TRUE
)
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Options please refer to \code{stats::p.adjust.methods}.}

\item{descending}{Logical (default: TRUE). Sort the ranks in descending order.}
}
\value{
A list with two vectors. \
\enumerate{
\item result: consists of descriptive statistics and compact letter display;
\item comparisons: includes statistics parameters for each pairwise comparisons.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
