% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_LSD_test.R
\name{Fisher_LSD_test}
\alias{Fisher_LSD_test}
\title{Fisher Least Significant Difference (Fisher-LSD) test}
\usage{
Fisher_LSD_test(data, formula, alpha = 0.05, p_adjust_method = "none")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Other options: \code{stats::p.adjust.methods}.}
}
\value{
A list with four vectors.
\enumerate{
\item tests: A message showing the statistical methods applied on the dataset.
\item pre_hoc: The result of pmnibus test.
\item post_hoc: includes statistics parameters for each pairwise comparisons.
\item cld: A dataframe reporting the descriptive stats and compact letter display.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\references{
Howell, D.C. 2013.
Statistical methods for psychology (8th ed.). pg. 393.
Wadsworth Cengage Learning, Belmont, CA.
}
\author{
Joon-Keat Lai
}
