% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Games_Howell_test.R
\name{Games_Howell_test}
\alias{Games_Howell_test}
\title{Games-Howell test}
\usage{
Games_Howell_test(data, formula, alpha = 0.05, p_adjust_method = "none")
}
\arguments{
\item{data}{A data frame.}

\item{formula}{A formula specifying the model, for example, yield ~ fertilizer_level.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Other options: \code{stats::p.adjust.methods}.}
}
\value{
A list with four elements:
\enumerate{
\item tests: A message showing the statistical methods applied on the dataset.
\item pre_hoc: The result of pmnibus test.
\item post_hoc: includes statistics parameters for each pairwise comparisons.
\item cld: A dataframe reporting the descriptive stats and compact letter display.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons (when group variances are unequal).
}
\examples{
set.seed(1)
df0 <- data.frame(
    group = rep(c("A", "B", "C"), each = 10),
    norm_data = c(rnorm(10, -1, 3), rnorm(10, 3, 1), rnorm(10, 0, 1)),
    skew_data = c(sqrt(rnorm(10, -7.2, 2) ^ 2), runif(10), runif(10))
)

out <- Games_Howell_test(df0, norm_data ~ group)
out$cld
#>   GROUP  N        AVG        SD          MED        MIN       MAX CLD
#> 1     B 10  3.2488450 1.0695148  3.491872279  0.7853001 4.5117812   a
#> 2     C 10 -0.1336732 0.9556076  0.009218122 -1.9893517 0.9189774   b
#> 3     A 10 -0.6033917 2.3417579 -0.230273356 -3.5068858 3.7858424   b
}
\references{
Howell, D.C. 2013.
Statistical methods for psychology (8th ed.). pg. 393.
Wadsworth Cengage Learning, Belmont, CA.
}
\author{
Joon-Keat Lai
}
