% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway_test.R
\name{oneway_test}
\alias{oneway_test}
\title{One-way test
Automatically performs one-way test with appropriate method, according to the data distribution
and homoscedasticity.}
\usage{
oneway_test(
  data,
  formula,
  alpha = 0.05,
  p_adjust_method = "holm",
  use_art = TRUE
)
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{Character string (default: "holm"). Other options: \code{stats::p.adjust.methods}.}

\item{use_art}{Whether using aligned rank transform. Default: TRUE}
}
\value{
A list, contains omnibus test and post-hoc test results.
}
\description{
One-way test
Automatically performs one-way test with appropriate method, according to the data distribution
and homoscedasticity.
}
\examples{
set.seed(1)
df0 <- data.frame(
    group = rep(c("A", "B", "C"), each = 10),
    norm_data = c(rnorm(10, -1, 2), rnorm(10, 3, 2), rnorm(10, 0, 1.5)),
    skew_data = c(sqrt(rnorm(10, -7.2, 2) ^ 2), runif(10), runif(10))
)

out <- oneway_test(df0, skew_data ~ group)
res <- out$result
res
#>   GROUPS  N       AVG        SD       MED        MIN       MAX CLD
#> A      A 10 6.9585396 1.6171293 7.3131184 4.48264090 9.9541191   a
#> C      C 10 0.5674507 0.2713306 0.6236108 0.05893438 0.8762692   b
#> B      B 10 0.4053906 0.2437230 0.3626733 0.12169192 0.7570871   b

out <- oneway_test(df0, norm_data ~ group)
res <- out$result
res
#>   GROUPS  N        AVG       SD         MED       MIN      MAX CLD
#> B      B 10  3.4976899 2.139030  3.98374456 -1.429400 6.023562   a
#> C      C 10 -0.2005099 1.433411  0.01382718 -2.984028 1.378466   b
#> A      A 10 -0.7355944 1.561172 -0.48684890 -2.671257 2.190562   b
}
