% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REGWQ_test.R
\name{REGWQ_test}
\alias{REGWQ_test}
\title{Ryan, Einot, Gabriel, Welsh Studentized Range Q (REGWQ) test}
\usage{
REGWQ_test(data, formula, alpha = 0.05, p_adjust_method = "none")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Options please refer to \code{stats::p.adjust.methods}.}
}
\value{
A list with two vectors. \
\enumerate{
\item result: consists of descriptive statistics and compact letter display;
\item comparisons: includes statistics parameters for each pairwise comparisons.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\examples{
rawdata <- data.frame(
    m1 = c(51, 84, 50, 48, 79, 61, 53, 54),
    m2 = c(82, 91, 92, 80, 52, 85, 73, 74),
    m3 = c(79, 84, 74, 98, 63, 83, 85, 58),
    m4 = c(85, 80, 65, 71, 67, 51, 63, 93),
    m5 = c(37, 40, 61, 51, 76, 55, 60, 70)
)

rawdata <- stats::reshape(
    data = rawdata,
    direction = "long",
    varying = colnames(rawdata),
    v.names = "val",
    times = colnames(rawdata),
    timevar = "group"
)

stats_res <- REGWQ_test(rawdata, val ~ group)
stats_res[["result"]]["CLD"]
#>    GROUPS N    AVG       SD  MED MIN MAX CLD
#> m2     m2 8 78.625 12.80555 81.0  52  92   a
#> m3     m3 8 78.000 12.82854 81.0  58  98   a
#> m4     m4 8 71.875 13.47418 69.0  51  93  ab
#> m1     m1 8 60.000 13.87701 53.5  48  84   b
#> m5     m5 8 56.250 13.51983 57.5  37  76   b

}
\references{
Howell, D.C. 2013.
Statistical methods for psychology (8th ed.). pg. 393.
Wadsworth Cengage Learning, Belmont, CA.
}
