% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey_HSD_test.R
\name{Tukey_HSD_test}
\alias{Tukey_HSD_test}
\title{Tukey Honestly Significant Difference (Tukey-HSD) test}
\usage{
Tukey_HSD_test(data, formula, alpha = 0.05, p_adjust_method = "none")
}
\arguments{
\item{data}{A data frame in which the variables specified in the formula will be found.}

\item{formula}{A formula specifying the model.}

\item{alpha}{Numeric value range from 0 to 1 (default: 0.05). The error tolerance.}

\item{p_adjust_method}{A character string (default: "none"). Other options: \code{stats::p.adjust.methods}.}
}
\value{
A list with two vectors.
\enumerate{
\item result: consists of descriptive statistics and compact letter display;
\item comparisons: includes statistics parameters for each pairwise comparisons.
}
}
\description{
Represent significance statements resulting from all-pairwise comparisons.
}
\references{
Howell, D.C. 2013.
Statistical methods for psychology (8th ed.). pg. 393.
Wadsworth Cengage Learning, Belmont, CA.

Tukey, John W. 1949.
Comparing Individual Means in the Analysis of Variance.
Biometrics, vol. 5, no. 2, pp. 99-114. JSTOR.
https://doi.org/10.2307/3001913.
}
